requirements.txt
html2text==2014.7.3
Markdown==2.4.1
nose==1.3.3
requests==2.6.0
PyYAML==5.1